+++ frontend/app/src/entities/proposed-changes/ui/proposed-changes-table.tsx
@@ -10,7 +10,7 @@ import { useFilters } from "@/entities/nodes/filters/ui/hooks/use-filters";
 import { ObjectTableEmpty } from "@/entities/nodes/object/ui/object-table/object-table-empty";
 import { useSort } from "@/entities/nodes/sort/ui/hooks/use-sort";
 import { computeProposedChangeFilters } from "@/entities/proposed-changes/domain/rules/compute-proposed-change-filters";
-import { computeProposedChangeSort } from "@/entities/proposed-changes/domain/rules/proposed-change-sort";
+import { computeProposedChangeSort } from "@/entities/proposed-changes/domain/rules/compute-proposed-change-sort";
 import { ProposedChangesItem } from "@/entities/proposed-changes/ui/proposed-change-item";
 import { ProposedChangesTableFilters } from "@/entities/proposed-changes/ui/proposed-changes-table-filters";
 import { ProposedChangesTableSkeleton } from "@/entities/proposed-changes/ui/proposed-changes-table-skeleton";
@@ -28,13 +28,11 @@ export function ProposedChangesTable({ schema, className }: ProposedChangesTable
   const [filters] = useFilters();
   const { appliedSort } = useSort(schema);
 
-  const sort = computeProposedChangeSort(appliedSort);
-
   const { data, fetchNextPage, hasNextPage, isPending, isFetchingNextPage } = useGetProposedChanges(
     {
       schema,
       filters: computeProposedChangeFilters({ filters, qsp: proposedChangeState as string }),
-      sort,
+      sort: computeProposedChangeSort(appliedSort),
     }
   );
